Add notification logic in phase 2 for pr < 20000

+++ KUBOnlinePRPM/Service/PRService.cs
@@ -9,7 +9,7 @@ namespace KUBOnlinePRPM.Service
     public class PRService
     {
         // check project amount
-        public bool CheckAmountScenarioOne(int amount)
+        public bool CheckAmountScenarioOne(decimal amount)
         {
             if (amount <= 20000)
                 return true;
@@ -17,7 +17,7 @@ namespace KUBOnlinePRPM.Service
             return false;
         }
 
-        public bool CheckAmountScenarioTwo(int amount)
+        public bool CheckAmountScenarioTwo(decimal amount)
         {
             if (amount >= 20001 && amount <= 50000)
                 return true;
@@ -25,7 +25,7 @@ namespace KUBOnlinePRPM.Service
             return false;
         }
 
-        public bool CheckAmountScenarioThree(int amount)
+        public bool CheckAmountScenarioThree(decimal amount)
         {
             if (amount >= 50001 && amount <= 300000)
                 return true;
